refactor(nodes): rename `BlockchainLabel` -> `NodeLabel`
