+++ src/components/nodes/adm/NodesTable/NodeLabel.vue
